requirements.txt
Discord.py>=1.3.1
Wavelink>=0.5.1
git+https://github.com/Rapptz/discord-ext-menus.git